+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.5.0
+
+### Added
+
+- `mrn_error_log.tsv`
+- Python scripts / jobs
+  - Curious data to REDCap
+    - `{activity_name}_start_date` to Curious instruments in REDCap
+
+### Fixed
+
+- bug in pushing Curous alerts to REDCap
+- bug in pushing REDCap participants to Curious
+
 ## 1.4.1
 
 ### Added